+++ R/netPath.R
@@ -11,59 +11,103 @@
 #'
 #' @examples
 #' \dontrun{
-#' make.Path(OD_Sample)
+#' make_net_path(OD_Sample)
 #'
 #' # Generate Leg Count
-#' make.Path(OD_Sample, leg = TRUE)
+#' make_net_path(OD_Sample, leg = TRUE)
 #' }
 #' @export
 #'
-#'
 
-make.Path <- function(x, leg = FALSE, zero = FALSE, carrier = FALSE){
+make_net_path <- function(x, leg = FALSE, zero = FALSE, carrier = FALSE){
 
-  print("This code might take longer than usual to execute")
+  message("This code might take longer than usual to execute")
 
   # Gets lasts values by Market and Sequence number
   DT <- as.data.table(x)
   DT <- DT[order(mkt_id, seq_num)]
 
-  if(zero == TRUE){
+  if(zero == TRUE & carrier == FALSE){
 
-    DT <- DT[, .(origin=origin[1], dest=dest[.N], itin_fare=itin_fare[1], passengers = passengers[1],
-                 roundtrip = roundtrip[1], itin_yield = itin_yield[1], num_stops = .N,
-                 pct_zero = ifelse(itin_fare == 0, 1, 0), distance = sum(distance)), by=mkt_id]
+    DT <- DT[, .(origin=origin[1], dest=dest[.N],
+                 itin_fare=itin_fare[1], passengers = passengers[1],
+                 roundtrip = roundtrip[1], itin_yield = itin_yield[1],
+                 num_stops = .N, pct_zero = ifelse(itin_fare == 0, 1, 0),
+                 distance = sum(distance), year, quarter), by=mkt_id]
 
     # Calculates averages and sums
     DT <- DT[, itin_fare := itin_fare/(1+roundtrip)]
-    netOD <- DT[, .(passengers = sum(passengers),fare_sd = round(sd(itin_fare), 2), itin_fare = round(sum(itin_fare)/(.N), 2),
-                    itin_yield = round(sum(itin_yield)/(.N), 3), mean_stops = round(sum(num_stops)/(.N)),
-                    pct_zero = round((sum(pct_zero)*100), 2)/(.N), mean_distance = round(sum(distance)/(.N), 2)), by=.(origin, dest)][order(origin, dest)]
+    netOD <- DT[, .(passengers = sum(passengers),
+                    fare_sd = round(sd(itin_fare), 2),
+                    itin_fare = round(sum(itin_fare)/(.N), 2),
+                    itin_yield = round(sum(itin_yield)/(.N), 3),
+                    mean_stops = round(sum(num_stops)/(.N)),
+                    pct_zero = round((sum(pct_zero)*100), 2)/(.N),
+                    mean_distance = round(sum(distance)/(.N), 2)),
+                by=.(origin, dest, year, quarter)][order(origin, dest)]
 
-  }else{
+  }
 
-    if(carrier == TRUE){
+  if(zero == TRUE & carrier == TRUE){
 
-      DT <- DT[, .(origin=origin[1], dest=dest[.N], itin_fare=itin_fare[1], passengers = passengers[1],
-                   roundtrip = roundtrip[1], itin_yield = itin_yield[1], op_carrier = op_carrier[1], num_stops = .N, distance = sum(distance)), by=mkt_id]
+    DT <- DT[, .(origin=origin[1], dest=dest[.N],
+                 itin_fare=itin_fare[1], passengers = passengers[1],
+                 roundtrip = roundtrip[1], itin_yield = itin_yield[1],
+                 op_carrier = op_carrier[1], num_stops = .N,
+                 pct_zero = ifelse(itin_fare == 0, 1, 0),
+                 distance = sum(distance), year, quarter), by=mkt_id]
+
+    # Calculates averages and sums
+    DT <- DT[, itin_fare := itin_fare/(1+roundtrip)]
+    netOD <- DT[, .(passengers = sum(passengers),
+                    fare_sd = round(sd(itin_fare), 2),
+                    itin_fare = round(sum(itin_fare)/(.N), 2),
+                    itin_yield = round(sum(itin_yield)/(.N), 3),
+                    mean_stops = round(sum(num_stops)/(.N)),
+                    pct_zero = round((sum(pct_zero)*100), 2)/(.N),
+                    mean_distance = round(sum(distance)/(.N), 2)),
+                by=.(origin, dest, op_carrier, year, quarter)][order(origin, dest, op_carrier)]
+
+  }
+
+    if(carrier == TRUE & zero == FALSE){
+
+      DT <- DT[, .(origin=origin[1], dest=dest[.N],
+                   itin_fare=itin_fare[1], passengers = passengers[1],
+                   roundtrip = roundtrip[1], itin_yield = itin_yield[1],
+                   op_carrier = op_carrier[1], num_stops = .N,
+                   distance = sum(distance), year, quarter), by=mkt_id]
 
       # Calculates averages and sums
       DT <- DT[, itin_fare := itin_fare/(1+roundtrip)]
-      netOD <- DT[, .(passengers = sum(passengers),fare_sd = round(sd(itin_fare), 2),
-                      itin_fare = round(sum(itin_fare)/(.N), 2), itin_yield = round(sum(itin_yield)/(.N), 3),
-                      mean_stops = round(sum(num_stops)/(.N)), mean_distance = round(sum(distance)/(.N), 2)), by=.(origin, dest, op_carrier)][order(origin, dest, op_carrier)]
+      netOD <- DT[, .(passengers = sum(passengers),
+                      fare_sd = round(sd(itin_fare), 2),
+                      itin_fare = round(sum(itin_fare)/(.N), 2),
+                      itin_yield = round(sum(itin_yield)/(.N), 3),
+                      mean_stops = round(sum(num_stops)/(.N)),
+                      mean_distance = round(sum(distance)/(.N), 2)),
+            by=.(origin, dest, op_carrier, year, quarter)][order(origin, dest, op_carrier)]
 
 
-    }else{
+    }
 
-      DT <- DT[, .(origin=origin[1], dest=dest[.N], itin_fare=itin_fare[1], passengers = passengers[1],
-                   roundtrip = roundtrip[1], itin_yield = itin_yield[1], num_stops = .N, distance = sum(distance)), by=mkt_id]
+  if(carrier == FALSE & zero == FALSE){
+
+      DT <- DT[, .(origin=origin[1],dest=dest[.N],
+                   itin_fare=itin_fare[1], passengers = passengers[1],
+                   roundtrip = roundtrip[1], itin_yield = itin_yield[1],
+                   num_stops = .N, distance = sum(distance),
+                   year, quarter), by=mkt_id]
 
       # Calculates averages and sums
       DT <- DT[, itin_fare := itin_fare/(1+roundtrip)]
-      netOD <- DT[, .(passengers = sum(passengers),fare_sd = round(sd(itin_fare), 2),
-                      itin_fare = round(sum(itin_fare)/(.N), 2), itin_yield = round(sum(itin_yield)/(.N), 3),
-                      mean_stops = round(sum(num_stops)/(.N)), mean_distance = round(sum(distance)/(.N), 2)), by=.(origin, dest)][order(origin, dest)]
+      netOD <- DT[, .(passengers = sum(passengers),
+                      fare_sd = round(sd(itin_fare), 2),
+                      itin_fare = round(sum(itin_fare)/(.N), 2),
+                      itin_yield = round(sum(itin_yield)/(.N), 3),
+                      mean_stops = round(sum(num_stops)/(.N)),
+                      mean_distance = round(sum(distance)/(.N), 2)),
+                  by=.(origin, dest, year, quarter)][order(origin, dest)]
 
     }
 
@@ -78,9 +122,11 @@ make.Path <- function(x, leg = FALSE, zero = FALSE, carrier = FALSE){
     netOD <- netOD %>%
       left_join(airportCode, by = "dest") %>%
       select(-latitude.x, -latitude.y, -longitude.x, -longitude.y) %>%
-      mutate(fare_sd = ifelse(is.na(fare_sd), 0, fare_sd))
+      mutate(fare_sd = ifelse(is.na(fare_sd), 0, fare_sd)) %>%
+      select(-year, -quarter, everything())
 
     if(leg == FALSE){
+
       return(netOD)
     }
 
@@ -109,11 +155,18 @@ make.Path <- function(x, leg = FALSE, zero = FALSE, carrier = FALSE){
       # Count words
       DT$legCount <- (stringr::str_count(DT$path, "\\S+"))-1
 
-      return(list(netOD = netOD, netLegCount = DT))
+      netlist <- list(netOD = netOD, netLegCount = DT)
+      class(netlist) <- "skynet"
+      return(netlist)
 
-    }
   }
 }
 
-globalVariables(c("mkt_id", "seq_num", "num_stops", "pct_zero", "latitude.y", "carriers",
-                  "description", ".", "carrier_name"))
+make.Path <- function(...){
+  warning(paste("make.Path is deprecated, use make_net_path(), instead."))
+  do.call(make_net_path, list(...))
+}
+
+globalVariables(c("mkt_id", "seq_num", "num_stops", "pct_zero",
+                  "latitude.y", "carriers", "description", ".",
+                  "carrier_name"))